Feat: 캐러셀 컴포넌트 분리
